Update core test files to use loadTestCalendar utility after pf2e extraction
Updated remaining core test files that were importing golarion-pf2e.json directly:

- calendar-engine-leap-year-regression.test.ts: Replaced direct import with loadTestCalendar
- year-calculation-behavior.test.ts: Updated calendar loading and PF2e setup import
- calendar-engine-worldtime-interpretation-regression.test.ts: Replaced direct import with loadTestCalendar
- widget-sync-integration.test.ts: Updated calendar loading
- calendar-engine-world-creation-timestamp.test.ts: Updated calendar loading and PF2e setup import
- time-converter-compatibility-manager.test.ts: Replaced direct import with loadTestCalendar

Fixed cross-package imports:
- Updated setup-pf2e imports to point to ../../pf2e-pack/test/setup-pf2e for core tests that need PF2e functionality

Fixed ESLint issues by removing unused imports and variables.

All files now use the loadTestCalendar utility which automatically finds calendars across multiple packages, making tests resilient to calendar location changes.
